+++ blazee/model.py
@@ -24,6 +24,7 @@ import io
 import logging
 import pickle
 import uuid
+import zipfile
 
 import requests
 from dateutil import parser
@@ -36,6 +37,17 @@ class ModelType(enum.Enum):
     SKLEARN = 'sklearn'
 
 
+def _generate_zip(file_name, content):
+    zipped = io.BytesIO()
+
+    with zipfile.ZipFile(zipped, mode="w", compression=zipfile.ZIP_DEFLATED) as zf:
+        zinfo = zipfile.ZipInfo(file_name)
+        zinfo.external_attr = 0o644 << 16  # give read access
+        zf.writestr(zinfo, content)
+
+    return zipped.getvalue()
+
+
 def _serialize_model(model):
     if _is_sklearn(model):
         if not type(model).__module__.startswith('sklearn.'):
@@ -46,7 +58,7 @@ def _serialize_model(model):
             for name, step in model.steps:
                 if not type(step).__module__.startswith('sklearn.'):
                     raise TypeError(
-                        f'Pipeline contains a step of type {type(step)} which is not supported. You must use a model or Pipeline that only contains estimators from the Scikit Learn library.')
+                        f'Pipeline contains step {name} of type {type(step)} which is not supported. You must use a model or Pipeline that only contains estimators from the Scikit Learn library.')
 
         # Check if model is fitted
         try:
@@ -56,9 +68,21 @@ def _serialize_model(model):
             if isinstance(e, NotFittedError):
                 raise AttributeError("This model hasn't been trained yet")
 
-        content = io.BytesIO()
-        pickle.dump(model, content)
-        return 'sklearn', content.getvalue()
+        content = pickle.dumps(model)
+        zipped = _generate_zip('model.pickle', content)
+        return 'sklearn', zipped
+    else:
+        raise TypeError(f'Model Type not supported: {type(model)}')
+
+
+def _get_model_metadata(model):
+    if _is_sklearn(model):
+        import sklearn
+        return {
+            'lib_versions': {
+                'scikit-learn': sklearn.__version__
+            }
+        }
     else:
         raise TypeError(f'Model Type not supported: {type(model)}')
 
@@ -77,22 +101,8 @@ class BlazeeModel:
 
     def __init__(self, client, response):
         self.client = client
-        self.id = response['id']
-        self.name = response['name']
-        if response['default_version']:
-            self.default_version = ModelVersion(client,
-                                                self,
-                                                response['default_version'])
-        else:
-            self.default_version = None
-        self.created_at = parser.parse(response['created_at'])
-        self.updated_at = parser.parse(response['updated_at'])
-
-    def __repr__(self):
-        return f"<BlazeeModel '{self.name}'\n\tid={self.id}>"
-
-    def __str__(self):
-        return f"<BlazeeModel '{self.name}'\n\tid={self.id}>"
+        self._deleted = False
+        self._reset(response)
 
     def rename(self, name):
         """Renames the model on Blazee
@@ -107,35 +117,46 @@ class BlazeeModel:
         self: `blazee.model.BlazeeModel`
             The model
         """
-        resp = self.client._api_call(f'/models/{self.id}',
-                                     method='PATCH',
-                                     json={
-                                         'name': name
-                                     })
-        self.name = resp['name']
+        self._check_deleted()
+        self.client._api_call(f'/models/{self.id}',
+                              method='PATCH',
+                              json={
+                                  'name': name
+                              })
+        self._refresh()
         return self
 
-    def get_version(self, version_id):
-        """Returns the version of this model with the given ID.
+    def versions(self):
+        """Returns all the version of this model
+
+        Returns
+        -------
+        versions: `list` of blazee.model.ModelVersion`
+        """
+        self._check_deleted()
+        resp = self.client._api_call(f'/models/{self.id}/versions')
+        return [ModelVersion(self.client, self, r) for r in resp]
+
+    def get_version(self, name_or_id):
+        """Returns the version of this model from its name or ID.
 
         Returns
         -------
         model_version: `blazee.model.ModelVersion`
         """
-        try:
-            uuid.UUID(version_id)
-        except ValueError:
-            raise ValueError(f'Malformed version ID: {version_id}')
+        self._check_deleted()
+        for version in self.versions():
+            if name_or_id in [version.id, version.name]:
+                return version
 
-        resp = self.client._api_call(
-            f'/models/{self.id}/versions/{version_id}')
-
-        return ModelVersion(self.client, self, resp)
+        raise ValueError(f"No version with ID or name {name_or_id}")
 
     def delete(self):
         """Deletes this model from Blazee"""
+        self._check_deleted()
         self.client._api_call(f'/models/{self.id}',
                               method='DELETE')
+        self._deleted = True
 
     def update(self, model, default=False):
         """Creates a new version of this model on Blazee and deploys it.
@@ -157,9 +178,11 @@ class BlazeeModel:
             The Blazee model that was deployed, ready to use for
             predictions
         """
+        self._check_deleted()
         model_type, content = _serialize_model(model)
+        metadata = _get_model_metadata(model)
 
-        version = self._upload_version(model_type, content)
+        version = self._upload_version(model_type, content, metadata)
         return version.model
 
     def predict(self, data):
@@ -180,6 +203,7 @@ class BlazeeModel:
         prediction: `blazee.prediction.Prediction`
             The prediction
         """
+        self._check_deleted()
         resp = self.client._api_call(f'/models/{self.id}/predict',
                                      method='POST',
                                      json=data)
@@ -205,17 +229,45 @@ class BlazeeModel:
         predictions: list of `blazee.prediction.Prediction`
             The predictions for the batch
         """
+        self._check_deleted()
         resp = self.client._api_call(f'/models/{self.id}/predict_batch',
                                      method='POST',
                                      json=data)
 
         return [Prediction(p) for p in resp]
 
-    def _upload_version(self, model_type, content):
+    def _reset(self, response):
+        self.id = response['id']
+        self.name = response['name']
+        if response['default_version']:
+            self.default_version = ModelVersion(self.client,
+                                                self,
+                                                response['default_version'])
+        else:
+            self.default_version = None
+        self.created_at = parser.parse(response['created_at'])
+        self.updated_at = parser.parse(response['updated_at'])
+
+    def _refresh(self):
+        resp = self.client._api_call(f'/models/{self.id}')
+        self._reset(resp)
+
+    def _check_deleted(self):
+        if self._deleted:
+            raise AssertionError('Model is deleted. Cannot perform operation')
+
+    def __repr__(self):
+        return f"<BlazeeModel '{self.name}'\n\tid={self.id}>"
+
+    def __str__(self):
+        return f"<BlazeeModel '{self.name}'\n\tid={self.id}>"
+
+    def _upload_version(self, model_type, content, metadata):
         resp = self.client._api_call(f'/models/{self.id}/versions',
                                      method="POST",
                                      json={
-                                         'type': model_type
+                                         'type': model_type,
+                                         'meta': metadata
                                      })
 
         upload_data = resp['upload_data']
@@ -231,6 +283,7 @@ class BlazeeModel:
             f'/models/{self.id}/versions/{version.id}/deploy',
             method='PATCH')
         logging.info(f"Successfully deployed model version {version.id}")
+        self._refresh()
         return ModelVersion(self.client, self, deploy_resp)
 
 
@@ -244,7 +297,7 @@ class ModelVersion:
         self.id = response['id']
         self.name = response['name']
         self.type = response['type']
-        self.data = response['data']
+        self.meta = response['meta']
         self.deployed = response['deployed']
         self.created_at = parser.parse(response['created_at'])
         self.updated_at = parser.parse(response['updated_at'])
@@ -308,6 +361,9 @@ class ModelVersion:
         Once this version is the default, it will be used to compute
         predictions for this model.
         """
-        self.client._api_call(f'/models/{self.model.id}/versions/{self.id}/deploy',
-                              method='PATCH')
-        self.model = self.client.get_model(self.model.id)
+        self.client._api_call(f'/models/{self.model.id}',
+                              method='PATCH',
+                              json={
+                                  'default_version_id': self.id
+                              })
+        self.model._refresh()